+++ dcpy/lifecycle/extract/templates/dcp_pad.yml
@@ -3,7 +3,7 @@ acl: public-read
 source:
   type: file_download
   url: https://s-media.nyc.gov/agencies/dcp/assets/files/zip/data-tools/bytes/pad_{{ version }}.zip
-transform_to_parquet_metadata:
+file_format:
   format: csv # TODO: this is not exactly correct. Need to think how to convert to parquet
 
 library_dataset:
