+++ dcpy/lifecycle/extract/templates/doe_pepmeetingurls.yml
@@ -4,7 +4,7 @@ source:
   type: script
   connector: dummy
   function: dummy
-transform_to_parquet_metadata:
+file_format:
   format: csv
 
 library_dataset:
