+++ dcpy/lifecycle/extract/templates/nysed_nonpublicenrollment.yml
@@ -3,7 +3,7 @@ acl: public-read
 source:
   type: file_download
   url: https://www.p12.nysed.gov/irs/statistics/nonpublic/enrollment-by-grade-nonpublic-{{ version }}.xlsx
-transform_to_parquet_metadata:
+file_format:
   format: xlsx
   tab_name: NonPubEnroll_byGrade_{{ version[2:] }} # TODO: need to confirm jinja works
