+++ dcpy/test/lifecycle/extract/resources/invalid_jinja.yml
@@ -3,4 +3,4 @@ acl: public-read
 source:
   type: file_download
   url: https://s-media.nyc.gov/agencies/dcp/assets/files/zip/{{ dummy_jinja_var }}/bytes/nyap_{{ version }}.zip
-transform_to_parquet_metadata:
+file_format:
